fixed paths, moved parsing of datetime to datetime_parser.py
